+++ src/api/routes/asset.router.ts
